app/backend/requirement.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2

llama-index==0.11.7
llama-index-llms-openai==0.2.9
llama-index-embeddings-openai==0.2.5
llama-index-readers-file==0.2.2
llama-index-vector-stores-faiss==0.2.1

pypdf==5.0.1
faiss-cpu==1.8.0.post1
